Modified
Modified the functionality of the animator to expect a complex scalar by default and added the possibility of choosing whether to plot the full complex order parameter or to just plot the director. Added colorbar label, as well as the option of having a time stamp to show the time progression of the animation.